feat:main 내부에 있던 mylotto분리
